libavb: Use size_t for lengths, not uint32_t.
Won't run into this problem with the way we use SHA-256 or SHA-512
(the partitions with hash descriptors - e.g. boot.img and others - are
typically less than 4 GiB) but it might be nice to fix anyway.  Bowgo
pointed this out a while ago but we never fixed it.

Bug: None
Test: All unit tests still pass (both on 32- and 64-bit).

+++ libavb/avb_sha512.c
@@ -318,9 +318,9 @@ static void SHA512_transform(AvbSHA512Ctx* ctx,
   }
 }
 
-void avb_sha512_update(AvbSHA512Ctx* ctx, const uint8_t* data, uint32_t len) {
-  unsigned int block_nb;
-  unsigned int new_len, rem_len, tmp_len;
+void avb_sha512_update(AvbSHA512Ctx* ctx, const uint8_t* data, size_t len) {
+  size_t block_nb;
+  size_t new_len, rem_len, tmp_len;
   const uint8_t* shifted_data;
 
   tmp_len = AVB_SHA512_BLOCK_SIZE - ctx->len;
@@ -350,9 +350,9 @@ void avb_sha512_update(AvbSHA512Ctx* ctx, const uint8_t* data, uint32_t len) {
 }
 
 uint8_t* avb_sha512_final(AvbSHA512Ctx* ctx) {
-  unsigned int block_nb;
-  unsigned int pm_len;
-  unsigned int len_b;
+  size_t block_nb;
+  size_t pm_len;
+  size_t len_b;
 
 #ifndef UNROLL_LOOPS_SHA512
   int i;
